TESTING: trying to get mesh info

+++ DRL_POL/scratch_visualize_qevent_witness.py
@@ -154,7 +154,7 @@ if __name__ == "__main__":
     df_last_timestep = Q_event_frames[1]
 
     # Convert the boolean Q values to integers for visualization
-    df_last_timestep["Q"] = df_last_timestep["Q"].astype(int)
+    df_last_timestep["Q"] = df_last_timestep["Q"].astype(float)
 
     # Parse the PVD file to extract mappings of timesteps to their corresponding PVTU files
     pvd_path = os.path.join(directory, pvdname)
